Progress on aprs
