--- codec2talkie/src/main/java/com/radio/codec2talkie/protocol/aprs/position/AprsPositionReport.java
@@ -1,8 +1,0 @@
-package com.radio.codec2talkie.protocol.aprs.position;
-
-public interface AprsPositionReport {
-    void fromPosition(AprsPosition position);
-    AprsPosition toPosition();
-    void fromBinary(byte[] data);
-    byte[] toBinary();
-}
