--- codec2talkie/src/main/java/com/radio/codec2talkie/protocol/aprs/position/AprsPositionReportTextWithTimestamp.java
@@ -1,24 +1,0 @@
-package com.radio.codec2talkie.protocol.aprs.position;
-
-public class AprsPositionReportTextWithTimestamp implements AprsPositionReport {
-
-    @Override
-    public void fromPosition(AprsPosition position) {
-
-    }
-
-    @Override
-    public AprsPosition toPosition() {
-        return null;
-    }
-
-    @Override
-    public void fromBinary(byte[] data) {
-
-    }
-
-    @Override
-    public byte[] toBinary() {
-        return new byte[0];
-    }
-}
